+++ src/game/game_fn_8008E430.c
@@ -10,7 +10,7 @@ typedef struct Data8008E430 {
 } Data8008E430;
 
 extern int fn_800DE3F8(void);
-extern u64 fn_8020123C();
+extern unsigned long long fn_8020123C();
 extern int fn_8008E078(void*, void*, void*);
 extern void fn_80038308(void*, int, short*);
 extern void fn_800389E0(void*, int, short, int);
