+++ src/game/game_fn_800CA6DC.c
@@ -20,7 +20,7 @@ typedef struct Event {
     EventArguments arguments;
 } Event;
 
-extern u32 fn_80201B54();
+extern int fn_80201B54();
 extern void fn_800CA660(void *, u32 *);
 extern int fn_800CA530(void *);
 extern int fn_801D0794(u32, int, u32, EventArguments *, int, int,
@@ -33,7 +33,7 @@ int fn_800CA6DC(void *object, int kind, u32 flags, void *value, int byte_value)
     u32 handle;
 
     event.flags = flags | kind;
-    handle = fn_80201B54(object);
+    handle = ((u32)fn_80201B54(object));
 
     switch (kind) {
     case 0x300:
